Add context to commands passed to parser
The parser needs to know which tab the user is currently on so that
it can execute the command with the right handler.

Let's add a new method in MainWindow.java that retrieves the current tab
the user is on, and modify the argument sent to parser by prepending
the tab name to the command entered by user.